separate the container karyawan into several small parts
