+++ src/pages/data-karyawan.tsx
@@ -1,8 +1,21 @@
+import { AdminLayout } from '@layouts/AdminLayout'
+import { KaryawanLayout } from '@layouts/KaryawanLayout'
 import { DataKaryawan } from '../containers/Karyawan'
+import { useMarContext } from 'src/context/MarFashionProvider'
 
 const DataKaryawanPage = () => {
+  const { user } = useMarContext()
   return (
-    <DataKaryawan />
+    <div>
+      { user?.role == "ADMIN" ? 
+        <AdminLayout>
+          <DataKaryawan />
+        </AdminLayout> :
+        <KaryawanLayout>
+          <DataKaryawan />
+        </KaryawanLayout>
+      }
+    </div>
   )
 }
